+++ VTS/VTS/Data/TriggerEvent.cs
@@ -11,7 +11,7 @@
         public string EventName { get; set; }
         public int Id { get; set; }
         public string ProxyMode { get; set; }
-        public double Radius { get; set; }
+        public float Radius { get; set; }
         public bool SphericalRadius { get; set; }
         public string TriggerMode { get; set; }
         public string TriggerType { get; set; }
